Update ShowRad.m
radiance conversion is still buggy.  I'm writing an email to Leif now for some help.  The raw image I extract is exactly the same as Convert2Tiff, so there's that at least.  You can uncomment the rad conversion (last steps in the m file) and see that it's pretty red.  Leif's is super blue, so I'm not sure if either is correct.  I hope we get some info from Leif soon.  There's room in the file to add dark subtraction by band.  I didn't have time to do that yet and was holding off to compare to Leif's convert2tiff global files.  We'll get there!  I'll work on this more later tonight.